+++ lib/Braintree/Transaction.php
@@ -169,6 +169,7 @@ namespace Braintree;
  * @property-read \Braintree\FacilitatedDetails $facilitatedDetails
  * @property-read \Braintree\FacilitatorDetails $facilitatorDetails
  * @property-read string $gatewayRejectionReason
+ * @property-read string $graphQLId transaction graphQLId
  * @property-read string $id transaction id
  * @property-read \Braintree\TransactionLineItem[] $lineItems
  * @property-read \Braintree\Transaction\MasterpassCardDetails $masterpassCardDetails transaction Masterpass card info